# optimize FFLexer.wantBytes() and FFLexer.lexNumber()