short story:
The idea behind the optimization is to replace multiple WriteByte(byte)
calls by single Write([]byte) call.

long story:
While profiling my application I noticed that some Buffer's functions
takes uncomfortable amount time. For example 10% for (*Buffer).grow().
Providing the lexer with preallocated buffer didn't change the picture.

```
...
2.90s  9.65% 34.31%      6.81s 22.66% github.com/pquerna/ffjson/fflib/v1.(*FFLexer).lexNumber
2.70s  8.99% 43.29%      2.88s  9.58% github.com/pquerna/ffjson/fflib/v1.(*Buffer).grow
2.60s  8.65% 51.95%      5.03s 16.74% github.com/pquerna/ffjson/fflib/v1.(*Buffer).WriteByte
2.16s  7.19% 59.13%     18.54s 61.70% github.com/pquerna/ffjson/fflib/v1.(*FFLexer).Scan
1.90s  6.32% 65.46%      1.90s  6.32% github.com/pquerna/ffjson/fflib/v1.(*ffReader).ReadByteNoWS
1.24s  4.13% 69.58%      1.24s  4.13% github.com/pquerna/ffjson/fflib/v1.(*Buffer).Reset
1.23s  4.09% 73.68%      1.23s  4.09% github.com/pquerna/ffjson/fflib/v1.scanString
1.15s  3.83% 77.50%      2.17s  7.22% github.com/pquerna/ffjson/fflib/v1.(*FFLexer).wantBytes
...
```

Further profiling and investigation showed that grow() is called for
each WriteByte() call. The later one is heavily used by lexer for
parsing a JSON document. But the hotspot was lexNumber() in my case
where WriteByte() is invoked for every character of a number (indeed,
the document I parsed contained thousands of unix timestamps which are
long numbers).

So, instead of calling WriteByte() after every parsed character my patch
makes the code to parse a number first and invoke Write([]byte) once for
the entire number. I also patched wantBytes() in similar fashion.

Profile after the change:

```
...
3.14s 10.45% 36.86%      4.53s 15.07% github.com/pquerna/ffjson/fflib/v1.(*FFLexer).lexNumber
2.53s  8.42% 45.28%     17.31s 57.58% github.com/pquerna/ffjson/fflib/v1.(*FFLexer).Scan
2.28s  7.58% 52.86%      2.28s  7.58% github.com/pquerna/ffjson/fflib/v1.(*ffReader).ReadByteNoWS
1.46s  4.86% 57.72%      1.46s  4.86% github.com/pquerna/ffjson/fflib/v1.scanString
1.22s  4.06% 61.78%      1.22s  4.06% github.com/pquerna/ffjson/fflib/v1.(*Buffer).Reset
1.10s  3.66% 65.44%      3.38s 11.24% github.com/pquerna/ffjson/fflib/v1.(*FFLexer).scanReadByte
1.08s  3.59% 69.03%      2.87s  9.55% github.com/pquerna/ffjson/fflib/v1.(*Buffer).Write
1.04s  3.46% 72.49%      1.25s  4.16% github.com/pquerna/ffjson/fflib/v1.(*Buffer).grow
...
```

Patched ffjson gave me 10-15% speed up.

Please note that some tests failed in master before the patch. So, I
didn't have nice test reports. But no new tests failed at least :-).